requirements.txt
PyYAML==6.0
Pillow==10.2.0
python-dateutil==2.8.2
matplotlib==3.8.2
